requirements.txt
discord.py==2.3.2
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
python-dotenv
gspread
oauth2client
python-dateutil
scipy